refactor user_terms_controller so that accept/reject code can be included elsewhere
